Renamed meta package to evidence to reflect what was in it.
