+++ src/main/java/org/phenopackets/api/model/association/PhenotypeAssociation.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.collect.ImmutableList;
 import org.phenopackets.api.model.condition.Phenotype;
 import org.phenopackets.api.model.entity.Entity;
-import org.phenopackets.api.model.meta.Evidence;
+import org.phenopackets.api.model.evidence.Evidence;
 
 import java.util.ArrayList;
 import java.util.List;
