+++ src/main/java/org/phenopackets/api/model/evidence/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author cjm
+ *
+ */
+package org.phenopackets.api.model.evidence;
