--- src/main/java/org/phenopackets/api/model/meta/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author cjm
- *
- */
-package org.phenopackets.api.model.meta;
